Make bottle code cross-platform.
