+++ Library/Homebrew/extend/os/bottles.rb
@@ -1,0 +1,5 @@
+require "utils/bottles"
+
+if OS.mac?
+  require "extend/os/mac/utils/bottles"
+end
