+++ Library/Homebrew/formula.rb
@@ -2,7 +2,7 @@ require "formula_support"
 require "formula_lock"
 require "formula_pin"
 require "hardware"
-require "bottles"
+require "utils/bottles"
 require "build_environment"
 require "build_options"
 require "formulary"
