+++ Library/Homebrew/formula_installer.rb
@@ -3,7 +3,7 @@ require "exceptions"
 require "formula"
 require "keg"
 require "tab"
-require "bottles"
+require "utils/bottles"
 require "caveats"
 require "cleaner"
 require "formula_cellar_checks"
