+++ Library/Homebrew/test/testball_bottle.rb
@@ -6,7 +6,7 @@ class TestballBottle < Formula
       stable.bottle do
         cellar :any_skip_relocation
         root_url "file://#{File.expand_path("..", __FILE__)}/bottles"
-        sha256 "9abc8ce779067e26556002c4ca6b9427b9874d25f0cafa7028e05b5c5c410cb4" => bottle_tag
+        sha256 "9abc8ce779067e26556002c4ca6b9427b9874d25f0cafa7028e05b5c5c410cb4" => Utils::Bottles.tag
       end
       cxxstdlib_check :skip
     end
