+++ Library/brew.sh
@@ -118,7 +118,9 @@ export HOMEBREW_LIBRARY
 export HOMEBREW_VERSION
 export HOMEBREW_CELLAR
 export HOMEBREW_RUBY_PATH
+export HOMEBREW_SYSTEM
 export HOMEBREW_CURL
+export HOMEBREW_PROCESSOR
 export HOMEBREW_PRODUCT
 export HOMEBREW_OS_VERSION
 export HOMEBREW_OSX_VERSION
